Move min version context to the version repo
